+++ src/snowflake/cli/plugins/nativeapp/commands.py
@@ -233,7 +233,7 @@ def app_deploy(
     files: Optional[List[Path]] = typer.Argument(
         default=None,
         show_default=False,
-        help="Paths of the files, relative to the deploy root, to be uploaded to a stage. [default: sync all local changes to the stage]",
+        help=f"""Paths of the files, relative to the deploy root, to be uploaded to a stage. If the path exists remotely but not locally, it will be deleted from the stage [default: sync all local changes to the stage]""",
     ),
     **options,
 ) -> CommandResult: